% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon2aa.R
\name{codon2aa}
\alias{codon2aa}
\title{codon2aa}
\usage{
codon2aa(data, display = FALSE)
}
\arguments{
\item{data}{Single caracter string of three characters, or vector of three characters, indicating the DNA codon (only "A", "T", "G" and "C" allowed). Case insensitive. Omitted if display argument is TRUE.}

\item{display}{Single logical value. Display the whole genetic table? if TRUE, override data.}
}
\value{
The 1 letter uppercase amino acid of the submitted codon or the whole table if display argument is TRUE.
}
\description{
Convert codon to amino acid using standard genetic code indicated in https://en.wikipedia.org/wiki/DNA_and_RNA_codon_tables.
}
\details{
REQUIRED PACKAGES

cuteDev


REQUIRED FUNCTIONS FROM THE cute PACKAGE

arg_check()


WARNINGS

None
}
\examples{
codon2aa(data = "ATC", display = TRUE)
}
