% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_finder.R
\name{codon_finder}
\alias{codon_finder}
\title{codon_finder}
\usage{
codon_finder(pos, begin, end)
}
\arguments{
\item{pos}{Vector of integers indicating the positions of nucleotids in a sequence. Must be between begin and end arguments.}

\item{begin}{Single integer indicating the position of the first base of the coding sequence.}

\item{end}{Single indicating the position of the last base of the coding sequence.}
}
\value{
A data frame with column names:
- pos: values of the pos argument.
- codon_nb: the codon number in the CDS encompassing the pos value.
- codon_pos: the position of pos in the codon (either 1, 2 or 3).
- codon_begin: the first base position of the codon.
- codon_end: the last base position of the codon.
}
\description{
Gives the codon number and position in the codon of nucleotid positions.
}
\details{
REQUIRED PACKAGES

cuteDev


WARNINGS

Only for coding sequences (no introns): ((end - begin) + 1) / 3 must be an integer (i.e., modulo zero)

Negatives positions allowed but this implies that one base has the position 0 in the sequence
}
\examples{
codon_finder(c(5, 6, 8, 10), begin = 5, end = 10)

codon_finder(c(0, 5, 6, 8, 10), begin = -2, end = 12)
}
