% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_finder.R
\name{codon_finder}
\alias{codon_finder}
\title{codon_finder}
\usage{
codon_finder(pos, begin, end, safer_check = TRUE)
}
\arguments{
\item{pos}{Vector of integers indicating the positions of nucleotids in a sequence. Must be between begin and end arguments.}

\item{begin}{Single integer indicating the position of the first base of the coding sequence.}

\item{end}{Single indicating the position of the last base of the coding sequence.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Must be set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
A data frame with column names:

- pos: values of the pos argument.

- codon_nb: the codon number in the CDS encompassing the pos value.

- codon_pos: the position of pos in the codon (either 1, 2 or 3).

- codon_begin: the first base position of the codon.

- codon_end: the last base position of the codon.
}
\description{
Gives the codon number and position in the codon of nucleotid positions.
}
\details{
WARNINGS

Only for coding sequences (no introns): ((end - begin) + 1) / 3 must be an integer (i.e., modulo zero)

Negatives positions allowed but this implies that one base has the position 0 in the sequence
}
\examples{
codon_finder(c(5, 6, 8, 10), begin = 5, end = 10)

codon_finder(c(0, 5, 6, 8, 10), begin = -2, end = 12)
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
